Renamed freediscovery.{api,server}
