+++ scripts/run_api_docker.py
@@ -6,6 +6,6 @@ from __future__ import print_function
 
 import sys
 
-from freediscovery.api import fd_app
+from freediscovery.server import fd_app
 
 app = fd_app('/freediscovery_shared/')
